Updated info related to spectral_ruleset
If you do not have a `.spectral.yaml` in the root of the project `spectral_ruleset` must be set. There is no such a thing as `the default built-in Spectral rulesets`.

Please note that this has changed with version `0.8.0` of the action since it uses version 6 of Spectral that changed the ruleset behavior https://github.com/stoplightio/spectral/issues/1796

I think that also the `Configuration` section should be removed because is just a repetition.